Don't offer several code actions on identifiers that are likely intended to be C# keywords.
