+++ src/Workspaces/CSharp/Portable/Extensions/SimpleNameSyntaxExtensions.cs
@@ -62,9 +62,17 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
                 return false;
             }
 
-            if (simpleName.IsVar)
+            if (simpleName.Identifier.CouldBeKeyword())
             {
-                // 'var' is much more likely to represent a keyword rather than "a possible type name".
+                // Something that looks like a keyword is almost certainly not intended to be a
+                // "Standalone type name".  
+                //
+                // 1. Users are not going to name types the same name as C# keywords (contextual or otherwise).
+                // 2. Types in .Net are virtually always start with a Uppercase. While keywords are lowercase)
+                //
+                // Having a lowercase identifier which matches a c# keyword is enough of a signal 
+                // to just not treat this as a standalone type name (even though for some identifiers
+                // it could be according to the language).
                 return false;
             }
